@@ -1,45 +1,58 @@
 /*****************************************************/
 //Organization: Stuba Green Team
-//Authors: Tereza Ábelová, Juraj Krasňanský
+//Authors: Tereza Ábelová, Juraj Krasňanský, Patrik Knaperek
 /*****************************************************/
 
-
-
 #include "../include/Messages.h"
+#include <ros/ros.h>
 #include <cmath>
 #include "opencv2/core/core.hpp"
 #include <sgtdv_msgs/Point2D.h>
+#include <sgtdv_msgs/Point2DArr.h>
+#include <sgtdv_msgs/CarPose.h>
+#include <sgtdv_msgs/CarVel.h>
+#include <visualization_msgs/Marker.h>
+#include <geometry_msgs/PoseStamped.h>
+
+constexpr float FPS = 60.f;
+constexpr float TIME_PER_FRAME = 1.f / FPS;
 
 #define deg2rad(x) (x*M_PI/180.f)
 #define rad2deg(x) (x*180.f/M_PI)
 
-constexpr float CAR_LENGTH = 2.f;
-constexpr float REAR_WHEELS_OFFSET = 0.9f;
-constexpr float FRONT_WHEELS_OFFSET = 1.1f;
-constexpr float CLOSEST_POINT_THRESHOLD = 0.2f;
-constexpr float CONTROL_GAIN = 1.f;
-
-
 class TrackingAlgorithm
 {
 protected:
-    TrackingAlgorithm();
-    ~TrackingAlgorithm();
+    TrackingAlgorithm(const ros::NodeHandle &handle);
+    ~TrackingAlgorithm() = default;
 
-    size_t m_coneIndexOffset;   //where to start looking for nearest point
+    virtual void VisualizePoint(const cv::Vec2f point, const int point_id, const cv::Vec3f color) const;
+    virtual void VisualizeSteering() const;
+    virtual void ComputeSpeedCommand(const float actSpeed);
+
+    ros::Publisher m_targetPub;
+    ros::Publisher m_steeringPosePub;
+    Control m_control;
+    Params m_params;
 
 public:
     virtual Control Do(const PathTrackingMsg &msg) = 0;
-    virtual void FreshTrajectory();
+    virtual void SetParams(const Params &params)
+    {
+        m_params = params;
+    };
+    virtual void SetVisualizationPublishers(const ros::Publisher &targetPub, const ros::Publisher &steeringPosePub)
+    {
+        m_targetPub = targetPub;
+        m_steeringPosePub = steeringPosePub;
+    };
 private:
 };
 
-
-
-class Stanley : public TrackingAlgorithm
+/*class Stanley : public TrackingAlgorithm
 {
 public:
-    Stanley();
+    Stanley(ros::NodeHandle &handle);
     ~Stanley();
 
     virtual Control Do(const PathTrackingMsg &msg);
@@ -50,19 +63,25 @@ private:
 
     void ComputeFrontWheelPos(const sgtdv_msgs::CarPose::ConstPtr &carPose);
     void ComputeThetaDelta(float theta, const cv::Vec2f &closestPoint);
-    cv::Vec2f FindClosestPoint(const sgtdv_msgs::Point2DArr::ConstPtr &trajectory);
+    cv::Vec2f FindTargetPoint(const sgtdv_msgs::Point2DArr::ConstPtr &trajectory);
     float ControlCommand(float speed) const;
     float SpeedGain(float speed) const;
-};
-
-
+};*/
 
 class PurePursuit : public TrackingAlgorithm
 {
 public:
-    PurePursuit();
-    ~PurePursuit();
+    PurePursuit(const ros::NodeHandle &handle);
+    ~PurePursuit() = default;
+
+    virtual Control Do(const PathTrackingMsg &msg);
 
 private:
+    cv::Vec2f m_rearWheelsPos;
+    float m_lookAheadDist;
 
+    void ComputeRearWheelPos(const sgtdv_msgs::CarPose::ConstPtr &carPose);
+    void ComputeLookAheadDist(const sgtdv_msgs::CarVel::ConstPtr &carVel);
+    cv::Vec2f FindTargetPoint(const sgtdv_msgs::Point2DArr::ConstPtr &trajectory) const;
+    void ComputeSteeringCommand(const PathTrackingMsg &msg, const cv::Vec2f &targetPoint);
 };
