@@ -1,26 +1,28 @@
 /*****************************************************/
 //Organization: Stuba Green Team
-//Authors: Juraj Krasňanský
+//Authors: Juraj Krasňanský, Patrik Knaperek
 /*****************************************************/
 
-
-#include <ros/ros.h>
 #include "../include/PathTrackingSynch.h"
-#include <sgtdv_msgs/Control.h>
 
 int main (int argc, char** argv)
 {
-    PathTrackingSynch synchObj;
-
     ros::init(argc, argv, "pathTracking");
     ros::NodeHandle handle;
+    PathTrackingSynch synchObj(handle);
 
-    ros::Publisher publisher = handle.advertise<sgtdv_msgs::Control>("pathtracking_commands", 1);
+    ros::Publisher commandPublisher = handle.advertise<sgtdv_msgs::Control>("pathtracking_commands", 1);
+    ros::Publisher targetPublisher = handle.advertise<visualization_msgs::Marker>("pathtracking_marker",1);
+    ros::Publisher steeringPosePublisher = handle.advertise<geometry_msgs::PoseStamped>("steering_angle_pathtracking", 1);
 
-    synchObj.SetPublisher(publisher);    
+    synchObj.SetCmdPublisher(commandPublisher);
+    synchObj.SetVisualizationPublishers(targetPublisher, steeringPosePublisher);    
 
     ros::Subscriber trajectorySub = handle.subscribe("pathplanning_trajectory", 1, &PathTrackingSynch::DoPlannedTrajectory, &synchObj);
-    ros::Subscriber actualStateSub = handle.subscribe("pose_estimate", 1, &PathTrackingSynch::DoPoseEstimate, &synchObj);
+    ros::Subscriber poseSub = handle.subscribe("pose_estimate", 1, &PathTrackingSynch::DoPoseEstimate, &synchObj);
+    ros::Subscriber velocitySub = handle.subscribe("velocity_estimate", 1, &PathTrackingSynch::DoVelocityEstimate, &synchObj);
+    ros::Subscriber stopSub = handle.subscribe("stop", 1, &PathTrackingSynch::StopCallback, &synchObj);
+    ros::Subscriber startSub = handle.subscribe("start", 1, &PathTrackingSynch::StartCallback, &synchObj);
 
     synchObj.Do();
 
